[refactor] #208 - 더 이상 사용되지 않는 Schedule 삭제 로직 추가, dailyJob 스케쥴러 시행 보장, repeatStartDate 계산 로직 핫픽스 (#211)
* [#208] refactor: 반복일정 수정/삭제 시 더 이상 사용되지 않는 반복일정이면 관련 일정 전체 삭제

* [#208] feat: 00시 스케쥴러에 연관된 scheduleDetail이 없는 단일일정 Schedule 삭제 로직 추가

* [#208] feat: 00시 스케쥴러에 연관된 scheduleDetail이 없는 단일일정 Schedule 삭제 로직 추가

* [#208] feat: redis를 사용해 dailyJob의 실행 보장

* [#208] hotfix: repeatStartDate가 제대로 계산되지 않았던 문제 해결

* [#208] refactor: schedule 삭제 시 연관된 discardedSchedule도 삭제

* [#208] refactor: 코드리뷰 반영

+++ src/main/java/com/kiero/schedule/application/service/ScheduleCommandService.java
@@ -106,7 +106,7 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 		if (request.isRecurring()) {
 
 			List<DayOfWeek> dayOfWeeks = dayOfWeekParser(request.dayOfWeek());
-			LocalDate repeatStartDate = repeatStartDateResolver(request.firstOrderDate(), dayOfWeeks, request.startTime(), today, now);
+			LocalDate repeatStartDate = repeatStartDateResolver(true, request.firstOrderDate(), dayOfWeeks, request.startTime(), today, now);
 
 			throwExceptionWhenAddRecurringScheduleIfDuplicated(request.dayOfWeek(), request.startTime(), request.endTime(), child.getId(), null, repeatStartDate, null);
 
@@ -134,7 +134,7 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 			// 3) 이벤트를 발행하도록 설정
 			DayOfWeek todayDayOfWeek = DayOfWeek.from(today.getDayOfWeek());
 
-			if (dayOfWeeks.contains(todayDayOfWeek) && request.startTime().isAfter(now) && !isFireLitToday && !isExistsTodayNotPendingAfterEndTime) {
+			if (repeatStartDate.isEqual(today) && dayOfWeeks.contains(todayDayOfWeek) && request.startTime().isAfter(now) && !isFireLitToday && !isExistsTodayNotPendingAfterEndTime) {
 
 				ScheduleDetail scheduleDetail = ScheduleDetail.create(today, null, null, ScheduleStatus.PENDING, null, saved);
 				scheduleDetailPersistencePort.save(scheduleDetail);
@@ -403,7 +403,7 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 
 				List<DayOfWeek> dayOfWeeks = dayOfWeekParser(request.dayOfWeek());
 
-				LocalDate repeatStartDate = repeatStartDateResolver(selectedDate, dayOfWeeks, request.startTime(), today, now);
+				LocalDate repeatStartDate = repeatStartDateResolver(false, selectedDate, dayOfWeeks, request.startTime(), today, now);
 
 				throwExceptionWhenAddRecurringScheduleIfDuplicated(request.dayOfWeek(), request.startTime(),
 					request.endTime(), childId, selectedDate, repeatStartDate, originalSchedule.getId());
@@ -434,7 +434,7 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 				// 1) 추가된 일정의 scheduleDetail 생성
 				// 2) 오늘 일정들의 stoneType 재계산
 				// 3) 이벤트를 발행하도록 설정
-				if (dayOfWeeks.contains(todayDayOfWeek) && request.startTime().isAfter(now) && !isFireLitToday && !isExistsTodayNotPendingAfterEndTime) {
+				if (repeatStartDate.isEqual(today) && dayOfWeeks.contains(todayDayOfWeek) && request.startTime().isAfter(now) && !isFireLitToday && !isExistsTodayNotPendingAfterEndTime) {
 
 					ScheduleDetail scheduleDetail = ScheduleDetail.create(today, null, null, ScheduleStatus.PENDING, null, saved);
 					scheduleDetailPersistencePort.save(scheduleDetail);
@@ -462,13 +462,18 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 					originalSchedule.getId());
 				List<DayOfWeek> requestDayOfWeeks = dayOfWeekParser(request.dayOfWeek());
 
-				LocalDate newScheduleRepeatStartDate = repeatStartDateResolver(selectedDate, requestDayOfWeeks, request.startTime(), today, now);
-				LocalDate originalScheduleRepeatEndDate = repeatEndDateResolver(selectedDate, originalDayOfWeeks);
+				LocalDate newScheduleRepeatStartDate = repeatStartDateResolver(false, selectedDate, requestDayOfWeeks, request.startTime(), today, now);
 
 				throwExceptionWhenAddRecurringScheduleIfDuplicated(request.dayOfWeek(), request.startTime(),
 					request.endTime(), childId, selectedDate, newScheduleRepeatStartDate, originalSchedule.getId());
 
-				originalSchedule.changeRepeatEndDate(originalScheduleRepeatEndDate);
+				LocalDate originalScheduleRepeatEndDate = repeatEndDateResolver(selectedDate, originalDayOfWeeks);
+
+				// 더 이상 유효하지 않은 반복 일정은 하드딜리트
+				if (originalScheduleRepeatEndDate.isBefore(originalSchedule.getRepeatStartDate())) {
+					deleteScheduleSet(originalSchedule);
+				}
+				else originalSchedule.changeRepeatEndDate(originalScheduleRepeatEndDate);
 
 				Schedule newSchedule = Schedule.create(
 					originalSchedule.getParent(),
@@ -497,7 +502,7 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 				// 2) 추가된 일정의 scheduleDetail 생성
 				// 3) 오늘 일정들의 stoneType 재계산
 				// 4) 이벤트를 발행하도록 설정
-				if (requestDayOfWeeks.contains(todayDayOfWeek) && request.startTime().isAfter(now) && !isFireLitToday && !isExistsTodayNotPendingAfterEndTime) {
+				if (newScheduleRepeatStartDate.isEqual(today) && requestDayOfWeeks.contains(todayDayOfWeek) && request.startTime().isAfter(now) && !isFireLitToday && !isExistsTodayNotPendingAfterEndTime) {
 					scheduleDetailPersistencePort.deleteByScheduleIdAndDate(originalSchedule.getId(), today);
 
 					ScheduleDetail scheduleDetail = ScheduleDetail.create(today, null, null, ScheduleStatus.PENDING, null, saved);
@@ -528,13 +533,22 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 				boolean sameDays = new HashSet<>(dayOfWeekParser(request.dayOfWeek())).equals(new HashSet<>(dayOfWeeks));
 				if (!sameDays) throw new KieroException(ScheduleErrorCode.DAY_OF_WEEK_CANNOT_BE_MODIFIED);
 
-				LocalDate newScheduleRepeatStartDate = repeatStartDateResolver(selectedDate, dayOfWeeks, request.startTime(), today, now);
-				LocalDate originalScheduleRepeatEndDate = repeatEndDateResolver(selectedDate, dayOfWeeks);
+				LocalDate newScheduleRepeatStartDate = repeatStartDateResolver(false, selectedDate, dayOfWeeks, request.startTime(), today, now);
 
 				throwExceptionWhenAddRecurringScheduleIfDuplicated(request.dayOfWeek(), request.startTime(),
 					request.endTime(), childId, selectedDate, newScheduleRepeatStartDate, originalSchedule.getId());
 
-				originalSchedule.changeRepeatEndDate(originalScheduleRepeatEndDate);
+				LocalDate originalScheduleRepeatEndDate = repeatEndDateResolver(selectedDate, dayOfWeeks);
+
+				// 원본 일정의 오늘자 scheduleDetail이 존재하는가 (= 오늘이 원본일정의 반복요일에 해당했는가)
+				boolean isOriginalDetailExists = scheduleDetailPersistencePort.findByScheduleIdAndDate(
+					originalSchedule.getId(), selectedDate).isPresent();
+
+				// 더 이상 유효하지 않은 반복 일정은 하드딜리트
+				if (originalScheduleRepeatEndDate.isBefore(originalSchedule.getRepeatStartDate())) {
+					deleteScheduleSet(originalSchedule);
+				}
+				else originalSchedule.changeRepeatEndDate(originalScheduleRepeatEndDate);
 
 				Schedule newSchedule = Schedule.create(
 					originalSchedule.getParent(),
@@ -556,11 +570,10 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 
 				scheduleRepeatDaysPersistencePort.saveAll(scheduleRepeatDays);
 
-				Optional<ScheduleDetail> originalDetail = scheduleDetailPersistencePort.findByScheduleIdAndDate(
-					originalSchedule.getId(), selectedDate);
-
-				if(originalDetail.isPresent()) {
-					originalDetail.get().changeSchedule(saved);
+				if(isOriginalDetailExists) {
+					ScheduleDetail scheduleDetail = ScheduleDetail.create(today, null, null, ScheduleStatus.PENDING,
+						null, saved);
+					scheduleDetailPersistencePort.save(scheduleDetail);
 					if (request.startTime().isAfter(now) && !isFireLitToday && !isExistsTodayNotPendingAfterEndTime) {
 						recalculateTodayStoneTypes(childId);
 						isEffectsToChildSchedule = true;
@@ -708,6 +721,16 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 					originalSchedule.getId());
 				LocalDate repeatEndDate = repeatEndDateResolver(selectedDate, dayOfWeeks);
 
+				// 더 이상 유효하지 않은 반복 일정은 하드딜리트
+				if (repeatEndDate.isBefore(originalSchedule.getRepeatStartDate())) {
+					deleteScheduleSet(originalSchedule);
+				}
+
+				// 1) 유효한 일정은 repeatEndDate 업데이트
+				// 삭제 요청한 일정이 오늘이고 유효하다면
+				// 2) scheduleDetail 삭제
+				// 3) 이벤트 발행
+				// 4) 불조각 종류 재계산
 				if ( !isToday ) {
 					originalSchedule.changeRepeatEndDate(repeatEndDate);
 				}
@@ -977,6 +1000,7 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 
 	// 반복일정이 첫 번째로 시작되는 일자를 구하는 리졸버
 	private LocalDate repeatStartDateResolver(
+		boolean isAdd,
 		LocalDate selectedDate,
 		List<DayOfWeek> repeatDays,
 		LocalTime startTime,
@@ -986,18 +1010,20 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 
 		LocalDate baseDate;
 
-		// 수정 후 반복 요일에 오늘이 포함되는지 여부
+		// 요청의 반복 요일에 오늘이 포함되는지 여부
 		boolean includesToday = repeatDays.contains(DayOfWeek.from(today.getDayOfWeek()));
 
 		if (selectedDate.isAfter(today)) { // 진입 일자가 미래 일자면
 
-			if (!includesToday) baseDate = selectedDate; // 반복요일에 오늘이 포함되지 않는다면
-			else if (startTime.isAfter(now)) baseDate = today; // 반복요일에 오늘이 포함되고, 일정 시작시간이 현재보다 이후라면
-			else baseDate = selectedDate; // 반복요일에 오늘이 포함되고, 일정 시작시간이 현재보다 이전이라면
+			if (isAdd) baseDate = selectedDate; // 일정 추가의 상황이라면
+			else if (includesToday && startTime.isAfter(now)) baseDate = today; // 일정 수정의 상황이고, 반복요일에 오늘이 포함되고, 일정 시작시간이 현재보다 이후라면
+			else baseDate = selectedDate; // 일정 수정의 상황이고, 반복요일에 오늘이 포함되고, 일정 시작시간이 현재보다 이전이라면
 
 		} else if (selectedDate.isBefore(today)) {
 
-			baseDate = today; // 진입 일자가 과거 일자면 baseDate는 오늘
+			// 반복 요일에 오늘이 포함되고, 일정 시작 시간이 현재보다 이후라면 baseDate는 오늘, 아니면 내일
+			if (includesToday && startTime.isAfter(now)) baseDate = today;
+			else baseDate = today.plusDays(1);
 
 		} else { // 진입 일자가 오늘이면
 			// 반복 요일에 오늘이 포함되고, 일자의 시작 시간이 현재 시간 이후라면 baseDate는 오늘
@@ -1028,5 +1054,12 @@ public class ScheduleCommandService implements ScheduleCommandUseCase {
 			cursor = cursor.minusDays(1);
 		}
 		return selectedDate.minusDays(1);
+	}
+
+	private void deleteScheduleSet(Schedule originalSchedule) {
+		scheduleDetailPersistencePort.deleteAllByScheduleId(originalSchedule.getId());
+		scheduleRepeatDaysPersistencePort.deleteAllByScheduleId(originalSchedule.getId());
+		schedulePersistencePort.deleteById(originalSchedule.getId());
+		discardedSchedulePersistencePort.deleteByScheduleId(originalSchedule.getId());
 	}
 }
